+++ fast_forward/__init__.py
@@ -210,6 +210,6 @@ python -m fast_forward.examples.interpolate \\
 __version__ = "0.1.0"
 
 from fast_forward.index import Mode
-from fast_forward.index.in_memory_index import InMemoryIndex
-from fast_forward.index.on_disk_index import OnDiskIndex
+from fast_forward.index.disk import OnDiskIndex
+from fast_forward.index.memory import InMemoryIndex
 from fast_forward.ranking import Ranking
