+++ fast_forward/index/memory.py
